Added the intial projects for the bot and some test code.
I added the bot console with a timer that gets the config file, reddit client, and then saves the wiki page on each run.

The initial check in for the config library is just populating the config from the app.config. It also grabs the wiki text from the wiki page, but doesn't do anything with it yet. It also includes a save config method, but this is unlikely to stay as the bot won't update it's own config.

The RedditApi library is a wrapper for the reddit sharp API. It abstracts the reddit sharp library to this simpler version. The reddit client handles all of the authentication and specific needs of the bot.