Added reddit sharp library to the solution because I had to make some modifications to the base code to get some features to work. Specfically, the SetFlair functionality. Hopefully I can remove the project later when it's fixed.
RedditApi - Added a RedditPost object so the main project can get a post object and interact with the underlying Post object without keeping that layer abstracted from the console. Changed the Reddit object instantiation to populate the bot user data. This is required for some checks later on in the process. I changed the UpdateSettings to wait for completion because the async nature of the method caused the program to close before the operation was completed. Changed the hard coded subreddit name to the shortcute in the config, minus the /r/. Added methods for: to get the last sitckied posts made by the bot, to get the latest posts from the sub's new queue,  to send a PM to a user, to get the bot's name, and to get a list of moderators from the sub.

Configuration - Added a private variable to store the sub shortcut so we could get other config data based on this. We wouldn't be able to get default config data otherwise. Added some null checks to the wiki config object so we can populate default values if the config is wiped out. Moved some code into a separate method to make it cleaner. Added a method to get the message to PM to users for free text friday. Added some additional setrtings for free text friday.

Bot - Removed dependency on the main config object. Completed FTF actions by completing the sitcky post method, adding PM users on new posts, and undoing everything on saturday.